allow run lengths >255bp

+++ src/MBGCommon.h
@@ -9,7 +9,7 @@
 using HashType = unsigned __int128;
 using NodeType = size_t;
 using CharType = uint16_t;
-using LengthType = uint8_t;
+using LengthType = uint16_t;
 using SequenceCharType = std::vector<CharType>;
 using SequenceLengthType = std::vector<LengthType>;
